requirements.txt
Django~=5.0.1
python-dotenv~=1.0.0
requests~=2.31.0